third_party/blink/web_tests/http/tests/xmlhttprequest/async-xhr-revalidate-after-sync-xhr-expected.txt

CONSOLE WARNING: Synchronous XMLHttpRequest on the main thread is deprecated because of its detrimental effects to the end user's experience. For more help, check https://xhr.spec.whatwg.org/.
Async XHR should not turn sync due to revalidation.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS xhr2SendIsReturned is true
PASS xhr1.responseText is not xhr2.responseText
PASS successfullyParsed is true

TEST COMPLETE
